Smaller PawnEntry. Bench: 3611854

+++ src/pawns.h
@@ -67,11 +67,11 @@ struct Entry {
   Bitboard pawnAttacksSpan[COLOR_NB];
   Square kingSquares[COLOR_NB];
   Score kingSafety[COLOR_NB];
-  int weakUnopposed[COLOR_NB];
-  int castlingRights[COLOR_NB];
+  int8_t weakUnopposed[COLOR_NB];
+  int8_t castlingRights[COLOR_NB];
   int semiopenFiles[COLOR_NB];
-  int pawnsOnSquares[COLOR_NB][COLOR_NB]; // [color][light/dark squares]
-  int asymmetry;
+  int8_t pawnsOnSquares[COLOR_NB][COLOR_NB]; // [color][light/dark squares]
+  int8_t asymmetry;
 };
 
 typedef HashTable<Entry, 16384> Table;
